requirements.txt
Django==4.0.6
django-webpack-loader
gunicorn==20.1.0
django-crispy-forms==1.14.0
Pillow==9.2.0
django-environ==0.9.0
psycopg2==2.9.3
djangorestframework==3.13.1
django-import-export==2.8.0
feedparser==6.0.10
selenium==4.4.3

celery==5.2.7
redis==4.3.4

django-celery-beat
django-celery-results

nltk==3.8.1
drf_writable_nested

django-countries-plus==2.1.0

langcodes==3.3.0
language_data

bs4

aspose-words==21.11.0 # convert docx to pdf (only Linux and Win)
python-docx # Reads, creates and writes docx
quickchart-io==1.0.1 #A Python client for the quickchart.io image charts web service.

rollbar==0.16.2

vaderSentiment==3.3.2
vader-multi==3.2.2.1 # multiple languages for vader

drf-yasg==1.21.4

fasttext-langdetect==1.0.5

geonamescache==1.5.0 # list of all cities
pycountry==22.3.5 # list of all countries
Scrapy==2.8.0
scrapy-djangoitem==1.1.1
google-search-results==2.4.1

wordcloud
django-ndarrayfield==1.0.0rc4 # model field for ML
sentence-transformers==2.2.2 # ML part
protobuf==3.19.0 # ML part

boolean-parser==0.1.4 # Expert Mode

deep-translator # translation
webdriver-manager==4.0.0

drf-multitokenauth

factory-boy==3.2.1

daphne==4.0.0
channels==4.0.0
drf-nested-routers==0.93.4
django-query-counter==0.2.2
django-phonenumber-field[phonenumberslite]
pandas==2.0.2
whatsapp-api-client-python==0.0.40
